Ingest a more accurate row count for smallish tables
Postgres' table statistics are updated regularly by autovacuum and are
good enough to get a feeling for how large tables are. However
autovacuum will only run if there have been sufficiently enough write
operations on that table to warrant the effort. Thus we missed small
changes to row counts.

Instead of relying on the table statistics, actually query those tables
and count the rows. However since on large tables this can generate a
lot of load and take a long time, we only do so on tables with less than
100 million rows. If a table is larger, presumably there is enough
traffic to get good numbers from PG's statistics anyway and it's less
important to know the exact number.