Cope with errors during agent collector auto-detection
Since we now have multiple collectors in the agent, we must not fail
completely if one of them returns an error. Indeed it's expected that
not all of them succeed in initializing. As long as at least one of them
succeeds, we're good to go. If not, show all the errors we encountered.